Added a Flash effect, which flashes a color on and off at a passed rate for a passed number of cycles.
This is used to show errors. I changed several sequences for reliability and added a Flash when one of a number of errors occurs. I also Flash green 3 times on startup to show the light is alive. When the flashing finishes it return -1 which destroys the Flash object and replaces it with a ConstantColor of 0. I also added a 500ms delay to startup which seems to make the startup flash more reliable. Finally, taking too long for a byte to come in used to just ignore the command. Now it generates and error flash.