Replaced Random with RandomFloat and RandomInt
They now take the same type of parameters as the return value. RandomFloat uses the random() function which works with int32_t. To get a floating point value I multiply the min and max by 1000, do random and then convert to float and divide by 1000. This reduces the range from 2 billion to 2 million, which should be fine.