Changed handling of newline
The previous newline handling in the Scanner was causing problems with comments at the end of lines. So I got rid of all that and now the Scanner just returns Token::NewLine when it sees like, like any other token. The CompileEngine now ignores newlines when they're not significant and expects them when they are.