Added support for Call/Return
New opcodes for Call and Return. Call has a target param which is 10 bit address relative to the start of code. A _codeOffset variable was added to get the absolute address of the Call. Upper 8 bits of the address are in the byte following the opcode, the lower 2 bits are in the lowest 2 bits of the opcode. So the opcode in in the 0x80 and above range. When Call is executed a 16 entry call stack saves the return pc address and then we jump to target + _codeOffset. When Return is executed, the previously pushed return address is popped and replaces _pc. All param passing and return values are by agreement between the caller and callee.