Changed Arly syntax
init and loop keywords are gone. At the end of the file is a lit of effects. As before they have the keyword effect, a single character name and the number of params. But added is an identifier of an init and loop function for the effect. All code is in the functions section. This also means that var definitions are at global scope, right after tables. This simplifies the engine because we don't need to worry about scope. And it lets functions share variables as needed. Not the most refined method, but it keeps everything simpler and will allow the more structured Clover language to be simpler to implement.