Redid all the indexed instructions
Got rid of LoadI and StoreI and renamed LoadX. This is a much cleaner system. To access a structure you do LoadRef to get the address of the base of the structure and then LoadDeref and StoreDeref to load and store values at offsets from that base. For Arrays, use LoadRefX. You pass a register with an index and a constant value with the number of words per array entry. The register value and constant are multiplied and added to the base address. Then LoadDeref and StoreDeref are used from there. The Arly example is updated and saves several bytes.