Added LoadTemp and StoreTemp
These are not yet tested. I'm adding them so I have a place to store temps as processing expressions at runtime. Right now there are slots for 4 temps with the address in the lower 2 bits of the op. But I don't think this will be enough so I'm moving the address to the next byte which will give room for 256 temps. But I will put the number of temps needed in the .arlx file and only allocate as much as needed at runtime