Redesigned variable handling
I've realized I need both global and local variables. Globals can be shared (for instance between init and loop). Locals allow you to save space because they are only active inside a function.

Now if an id is < 0x80 it is a constant (const or table), 0x80 to 0xbf is a global var (in the _global array) and from 0xc0 to 0xff is a local, relative to the _bp for the current function. Locals are not being generated yet, so they are untested. But the code to handle them is there. Functions now use a _bp to identify the start of local storage. It is loaded and restored on function call and return, but it is untested. The .arly test mostly works. But for some reason two of the flicker LEDs don't light up. I will investigate that soon