WIP - Started handling native functions
Lots of restructuring. Now Symbol and Function are classes. All dealing with addresses of the different Storage classes are handed inside Symbol. Function now contains a list of formal params and can hold a NativeFunction.A NativeFunction is simply an indicator that it is one and an address that is an enumerator for all the currently handled native functions. The list is in the Interpreter. There will be a new CallNative opcode which have this enumerator as its param so the Interpreter can call the function. None of this is generated, but it's all parsed and stored properly.