Dealing with types better
Type checking needs to match types. But handing pointers the type checking doesn't work. So added a Ptr type. These must match too, but it's not checking the underlying type. So you can have a pointer to anything and it will pass.
'c' command now works fully and the 'p' command works, but seems to have a Val that is too dark. Haven't tested 'f' command yet