Added annotation support
An annotation is a string of one line of code and a 32 bit integer. The integer is either -1, meaning this line doesn't have an address associated with it. Or it is an address in the code. When the decompiler runs it first outputs all the annotation lines with -1. Then it waits until it hits the line of code matching the next annotation integer. Then it outputs that line and all subsequent lines with -1. This interleaves the source code approximately with the generated code.