Implemented pre and post increment and decrement
I had originally decided not to do these. But the code to increment by hand is expensive and I found a way to do this efficiently. For both pre and post I push a ref on the stack and then the ops use that ref to get the value, increment or decrement it, store the result and then push either the original or modified value based on whether it is pre or post.