Made for loops nestable.
Got rid of ForEach and EndForEach. Replaced them with Loop (jump backwards) and Jump (jump forward). A for loop now has an if statement at the top which tests the variable against the end value. If >= it does a break, which jumps to the end of the loop. At the end of the for statement is an increment of the variable followed by a Loop instruction back to the beginning. No more member variables holding values for the loop, so loops can be nested. The engine supports the concept of 'break' and 'continue' but those statements are not implemented yet.