Added support for non-segmented binary files and .h formatted data files
You can now write binary output as a single file. You can also output a .h file with the binary data represented as hex bytes. This will all a sketch to be written that will compile in the binary so it can be copied to EEPROM. That way you don't need to upload over a serial line