Added support for talking to all lights on all posts
The ESP version will be a single ESP32 with 8 pixel rings on each post, all daisy chained together. This will talk to each group of 8 so each post will still operate indepdently

Mostly works. There's a bug in multicolor. It flickers on start and one of the colors is blank.